+++ CHANGELOG.md
@@ -67,6 +67,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - sdk(sdk): Added support for `set_should_keep_position_account` instruction.
 - sdk(sdk): Added support for `update_fees_state` instruction.
 - cli: Added `lp init-lp` command.
+- cli: Added `exchange close-empty-positions` command.
 
 ### Changed